refactor(test): deduplicate streaming test helpers into shared modules
- Extract compare_or_known into streaming_compare_support.rs shared by
  high-risk-structures and specific-regression test files
- Move normalize_whitespace_tokens and diff_summary into
  streaming_test_support.rs as canonical shared helpers
- Simplify known-difference matching (treat all matches as acceptable)
- Improve unified_diff_summary to report up to 3 differing lines
- Fix read_fixture_meta to probe both legacy and direct .meta.json paths
- Extract PEAK_VS_INPUT_THRESHOLD constant in evidence pack test
- Fix chunk-size calculation in high-risk-structures half-split
